// src/modules/inventory/reports/yarn-store/yarn-return-challan-report/yarn-challan-report-type.ts
export interface YarnReturnChallanReportType {
  COMPANY_NAME: string;
  COMPANY_ADDRESS: string;
  COM_CONTACT: string;
  CHALLAN_DATE: string;
  CHALLAN_NO: string;
  LC_NUMBER: string;
  WORK_ORDER_NUMBER: string;
  YARN_SUPPLIER: string;
  SUPPLIER_ADDRESS: string;
  YARN: string;
  YARN_BRAND: string;
  YARN_LOT_NUMBER: string;
  CTN: number;
  QTY: number;
  CONE: number;
  RETURN_PER_CTN_QTY: number;
  UOM: string;
  REMARKS: string;
  MAIN_COMPANY: string;
  MAIN_COMPANY_ADDRESS: string;
}